Move media browser code to dedicated file
